SpaceConsistency: Move to misc
It belongs to the other generic bears.
