+++ bears/misc/SpaceConsistencyBear.py
