--- bears/spacing/__init__.py
@@ -1,3 +1,0 @@
-"""
-Contains spacing related bears.
-"""
