+++ bears/tests/misc/SpaceConsistencyBearTest.py
@@ -1,12 +1,12 @@
-from queue import Queue
 import sys
+from queue import Queue
 
 sys.path.insert(0, ".")
 import unittest
 from coalib.settings.Setting import Setting
 from bears.tests.LocalBearTestHelper import LocalBearTestHelper
-from bears.spacing.SpaceConsistencyBear import (SpaceConsistencyBear,
-                                                SpacingHelper)
+from bears.misc.SpaceConsistencyBear import (SpaceConsistencyBear,
+                                             SpacingHelper)
 from coalib.settings.Section import Section
